ecommerce-data-platform/requirements-dev.txt
# ─── Dev dependencies for local IDE resolution ───────────────────────────
# Install these in a local venv so your IDE stops flagging import warnings.
# The full production stack runs inside Docker (see requirements.txt).
#
# Usage:
#   python -m venv .venv
#   .venv\Scripts\activate        (Windows)
#   pip install -r requirements-dev.txt

#  Core 
pydantic==2.7.1
python-dotenv==1.0.1
tenacity==8.3.0
faker==25.2.0

#  Kafka 
confluent-kafka==2.4.0

#  Database 
asyncpg==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30

# API 
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0

# Dashboard 
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
requests==2.32.0

# Testing 
pytest-asyncio==0.23.7
pytest-mock==3.14.0
pytest-cov==5.0.0

# ─── Spark (stub — install full PySpark for local Spark tests) ───────────
pyspark==3.5.1
pyarrow==16.0.0
